allowed embeds to recieve params
